remove alternate TradeUI
The only thing this did was making trade UI categories collapsible. However, due to constant updates by the game devs in this menu, it has repeatedly blocked/broken other features and led to negative user feedback, so it feels like it's time to retire this from the mod. It was originally hoped that this change would be implemented in the base game, but that never happened.